Collision sistem reworked and cleanning the code
The collision sistem now has his own method that can be called.
The global variable player that was in the Enemy class method to see the player score was removed. Now the engine do the wok with the methods to see and change the score/position of the objects.

Was removed the old commented code.